Add Gherkin Metrics node project
This project was meant to replace a Python stack. It generates data regarding all Gherkin in a defined repository to have a HTML view of the test debt per repository.
It was meant to be usable by QA Analysts on the project.

The original project could generate CSV files as a HTML data table of the test debt.

Most of the data has been stripped out for NDA issues, but this is to give the gross idea of the project.