Gherkin Doc Generator
Based of the Gherkin-Metrics project, this project extracts Gherkin Scenario to generate a readable documentation for those less familiar with repositories. It was also made to allow QA Analyst to easily search other projects without the need to clone every single repositories.

The HTML page generated is a simple Springboot interface.